pyproject.toml
[tool.poetry]
name = "tasos-apiauth"
version = "0.1.0"
description = "A re-usable FastAPI library that implements authentication, users, groups and permission handling."
authors = ["tristeng <tristen.georgiou@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [
    {include = "tasos"}
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.1"
python-multipart = "0.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
pydantic = {extras = ["email", "dotenv"], version = "^2.6.0"}
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"
pytest-asyncio = "^0.23.4"
httpx = "^0.26.0"
aiosqlite = "^0.19.0"
mypy = "^1.8.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.25"}
flake8 = "^7.0.0"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"
black = "^24.1.1"
uvicorn = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
env = [
    "APIAUTH_SECRET_KEY=somesecretkey",
    "APIAUTH_ALGORITHM=HS256",
    "APIAUTH_ACCESS_TOKEN_EXPIRE_MINUTES=15",
    "APIAUTH_DATABASE_URL=sqlite+aiosqlite://?check_same_thread=false",
]
markers = [
    "integration: mark test as an integration test.",
]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*:",
    "ignore::DeprecationWarning:pydantic.*:",
    "ignore::DeprecationWarning:jose.*:",
]

[tool.coverage.run]
command_line = "-m pytest tests -m 'not integration'"
branch = true
source = [
    "tasos",
]

[tool.coverage.report]
omit = [
    "tasos/apiauth/api/*",  # this file is tested with the fastapi test client, so coverage isn't reported
]
show_missing = true

[tool.mypy]
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin",
]
packages = [
    "tasos",
    "tests",
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true